requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
-e git+https://github.com/makinacorpus/django-leaflet.git@6805f167e08fcf2957993ac10ff340d2eeb5c9e3#egg=django_leaflet
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.8
